Optimization on plane allocation
BZ: 190141

1, allocate overlay to video layer whenever possible
2, allocate overlay A to pipe A and overlay C to pipe C whenever possible
3, allocate as many sprite planes as possible
4, plane allocation happens only when all conditions are met.
5, use rotation buffer provider to produce rotated buffer when needed

+++ ips/tangier/TngDisplayContext.h
@@ -53,7 +53,7 @@ private:
     IMG_display_device_public_t *mIMGDisplayDevice;
     IMG_hwc_layer_t mImgLayers[MAXIMUM_LAYER_NUMBER];
     bool mInitialized;
-    int mCount;
+    size_t mCount;
 };
 
 } // namespace intel
